requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.1.0  # or PyMySQL==1.0.3
transformers==4.33.3
torch==2.5.1
python-dotenv==1.0.0  # optional if you want .env support
